+++ front/src/components/Gainburua/components/KontaktuDatuak/KontaktuDatuak.module.scss
@@ -1,4 +1,4 @@
-@import '../../styles/tools/mixins/font.scss';
+@import '../../../../styles/tools/mixins/font.scss';
 
 .wrapper {
   display: flex;
