+++ front/src/components/Gainburua/components/KontaktuDatuak/index.ts
@@ -1,0 +1,1 @@
+export { KontaktuDatuak } from './KontaktuDatuak';
